+++ app/src/main/java/team/retum/jobisandroidv2/root/RootScreen.kt
@@ -12,7 +12,7 @@ import team.retum.bookmark.navigation.bookmark
 import team.retum.home.navigation.NAVIGATION_HOME
 import team.retum.home.navigation.home
 import team.retum.jobisandroidv2.ui.BottomNavigationBar
-import team.retum.recruitment.recruitments
+import team.retum.recruitment.navigation.recruitments
 
 @Composable
 internal fun Root(
