+++ app/src/main/java/team/retum/jobisandroidv2/ui/BottomMenu.kt
@@ -4,7 +4,7 @@ import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import team.retum.home.navigation.NAVIGATION_HOME
 import team.retum.jobisandroidv2.R
-import team.retum.recruitment.NAVIGATION_RECRUITMENTS
+import team.retum.recruitment.navigation.NAVIGATION_RECRUITMENTS
 
 sealed class BottomMenu(
     val route: String,
